+++ StudyGamePrograming03_VB/CircleComponent.vb
@@ -8,7 +8,7 @@ Public Class CircleComponent
     Sub New(ByRef owner As Actor, ByVal updateOrder As Integer)
         MyBase.New(owner, updateOrder)
     End Sub
-    Public Function GetRadius() As Double
+    Public Function GetRadius() As Single
         Return mOwner.GetRadius()
     End Function
 
@@ -20,10 +20,10 @@ End Class
 Public Module CircleModule
     Public Function Intersect(ByRef a As CircleComponent, ByRef b As CircleComponent) As Boolean
         ' ２つのCircleComponentの中心間距離を計算
-        Dim distSq As Double = (a.GetCenter() - b.GetCenter()).LengthSquared
+        Dim distSq As Single = (a.GetCenter() - b.GetCenter()).LengthSquared
 
         ' ２つのCircleComponentの半径の和を計算 
-        Dim sumRadiusSq As Double = (a.GetRadius() + b.GetRadius()) * (a.GetRadius() + b.GetRadius())
+        Dim sumRadiusSq As Single = (a.GetRadius() + b.GetRadius()) * (a.GetRadius() + b.GetRadius())
 
         ' 中心間距離 <= 半径の和 のとき、衝突したと判定
         If distSq <= sumRadiusSq Then
